- Many, many cleanups.   - Remove unused variables and functions   - Use function prototypes and declare missing function parameters   - Correct return values   - Initialise variables that might otherwise be used without initial values   - Align types (no declaring integers as pointers any more)   - Coerce 64-bit values into 32-bit ones for printing purposes (the only reason to use 64-bit values was to prevent problems     with the hashtable pointer sizes being different from the integer size. We don’t need the extra storage space)   - Deal with incoming values (sscanf, fscanf) properly by importing into a 32-bit int and then setting the 64-bit value   - Coerce types before passing to system library functions so that they match   - Remove redundant code that cannot be executed.   - Wrap if(assign) into if ((assign)) to prevent compiler warnings. At least a couple of these were bugs   - re-type functions that take floats to take doubles, since K&R does an auto-promote of a float to a double

+++ src/Ylib/stat.c
@@ -60,7 +60,7 @@ REVISIONS:  Sun Apr 21 21:21:58 EDT 1991 - renamed to Ystat_
 static char SccsId[] = "@(#) stat.c (Yale) version 3.3 4/21/91" ;
 #endif
 
-#include <yalecad/base.h>
+#include "yalecad/base.h"
 
 
 /* Given an array of number_elements of size_element, calculate min */
@@ -108,8 +108,7 @@ INT  size_element ;
 	    }
 	}
 	break ;
-    default: fprintf( stderr, "Unsupported element size:%d\n",
-	size_element ) ;
+    default: fprintf( stderr, "Unsupported element size:%d\n", (int)size_element ) ;
     }
 
     return(min);
@@ -160,8 +159,7 @@ INT  size_element ;
 	    }
 	}
 	break ;
-    default: fprintf( stderr, "Unsupported element size:%d\n",
-	size_element ) ;
+    default: fprintf( stderr, "Unsupported element size:%d\n", (int)size_element ) ;
 	
     }
     return(max);
@@ -204,8 +202,7 @@ INT  size_element ;
 	    sum += (DOUBLE) darray[i] ;
 	}
 	break ;
-    default: fprintf( stderr, "ERROR:Unsupported element size:%d\n",
-	size_element ) ;
+    default: fprintf( stderr, "ERROR:Unsupported element size:%d\n", (int)size_element ) ;
     }
 
     if( number_elements ){
@@ -264,8 +261,7 @@ DOUBLE mean ;
 	    sum2 += sum * sum ;
 	}
 	break ;
-    default: fprintf( stderr, "ERROR:Unsupported element size:%d\n",
-	size_element ) ;
+    default: fprintf( stderr, "ERROR:Unsupported element size:%d\n", (int)size_element ) ;
 
     }
     if( number_elements > 1 ){
